feature: lazy-load sibling Objective-C DLLs on Win32 missing class/method lookups
* Windows runtimes can now attempt to load nearby Objective-C DLLs when a class or method can’t be resolved (primarily to simplify test executables)
* non-delegating infraclass lookup now uses the fast-class table when available, avoiding side effects during dependency/loadinfo resolution
* **BREAKING**: `_mulle_objc_universe` callback wiring changed (`callbacks` → `callback`; `class_is_missing`/`method_is_missing` now return “nochange” status to control retry)

Introduced Win32-only `_mulle_objc_universe_dll_loader[_filtered]` and wired missing-class/method callbacks into class cache refresh and method search retry loops.

This DLL auto-loading is meant as a convenience for the test setup; production code should explicitly preload the desired DLL set (e.g. in `main`) before the first Objective-C call.

+++ src/mulle-objc-method.c
@@ -85,7 +85,7 @@ int  _mulle_objc_descriptor_pointer_compare_r( void **_a, void **_b, void *thunk
    struct _mulle_objc_descriptor   *a = * (struct _mulle_objc_descriptor **) _a;
    struct _mulle_objc_descriptor   *b = * (struct _mulle_objc_descriptor **) _b;
    int                             cmp;
-   long                            diff;
+   int32_t                         diff;
 
    MULLE_C_UNUSED( thunk);
 
@@ -97,7 +97,7 @@ int  _mulle_objc_descriptor_pointer_compare_r( void **_a, void **_b, void *thunk
    if( cmp)
       return( cmp);
 
-   diff = (long) a->bits - (long) b->bits;
+   diff = (int32_t) a->bits - (int32_t) b->bits;
    return( diff < 0 ? -1 : ! ! diff);
 }
